Add a news feature reflecting the casperjs news test.
Mink doesn't allow us to select multiple elements into an array like
javascript does, so we can't implement the test as fully as in the
casper test. However, this test is a bit overbearing anyway and we're
already asserting the truth in the behat version.

Remaining for UA-DIGITAL 265: Create instructions!