Create an easy ordering view for content managers.
Add the draggableviews module as a dependency, per @macaulay's
suggestion.

Using draggableviews, create a new view for reordering slides. It's
linked to by the contextual links in the Hero Carousel and is available
at `/admin/reorder-slides`. It requires the `Edit any Carousel Item`
permission to access.

Change the sort order for the Hero Carousel view to use the weight
created by draggableviews.