Incorporate feedback into reordering carousel feature.
Instances of the name of this feature have been changed to use the word
'carousel' instead of 'slides' (to avoid confusion) and to use 'uaqs' to
specify for site managers where the view is coming from.

Additional fields have been added to the view, so that the content
manager can more easily distinguish carousel items. These include a
thumbnail of the image and the call to action.